+++ src/DTO/SimpleStat.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App\DTO;
+
+class SimpleStat extends AbstractStat
+{
+}