+++ src/Repository/VariantStatRepository.php
@@ -12,6 +12,7 @@ use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
 
 /**
+ * @deprecated https://github.com/VEAF/website/issues/365
  * @method VariantStat|null find($id, $lockMode = null, $lockVersion = null)
  * @method VariantStat|null findOneBy(array $criteria, array $orderBy = null)
  * @method VariantStat[]    findAll()